Added enemy movement
Enemies move from the left of the screen to the right of the screen. There is always a fixed number of enemies alive. When an enemy makes it all the way to the right of the screen, its position is reset to the left of the screne, in a random stone row, and its speed is changed to a random value within a given range.